+++ src/quantum/quantum_context.h
@@ -148,7 +148,7 @@ public:
     int closeBuffer();
     int getNumCoroutineThreads() const;
     int getNumIoThreads() const;
-    const std::pair<size_t, size_t>& getCoroQueueIdRangeForAny() const;
+    const std::pair<int, int>& getCoroQueueIdRangeForAny() const;
 
     
     //===================================
